♻️ refactor(User): 사용 가능 닉네임 체크 모듈 위치 변경
