+++ src/main/java/com/wccg/well_c_code_git_backend/domain/user/controller/UserController.java
@@ -33,7 +33,7 @@ public class UserController {
     )
     public ResponseEntity<ApiResponse<NicknameAvaliableCheckResponse>> nicknameAvailableCheck(@RequestParam String nickname) {
 
-        ServiceNicknameAvailableCheckResponse serviceResponse = userService.nicknameAvailableCheck(toServiceNicknameAvailableCheckRequest(nickname));
+        ServiceNicknameAvailableCheckResponse serviceResponse = userProfileService.nicknameAvailableCheck(toServiceNicknameAvailableCheckRequest(nickname));
 
         return ResponseEntity
                 .status(HttpStatus.OK)
