+++ src/main/java/com/wccg/well_c_code_git_backend/domain/user/service/UserService.java
@@ -1,18 +1,12 @@
 package com.wccg.well_c_code_git_backend.domain.user.service;
 
-import com.wccg.well_c_code_git_backend.domain.user.dto.service.request.ServiceNicknameAvailableCheckRequest;
-import com.wccg.well_c_code_git_backend.domain.user.dto.service.response.ServiceNicknameAvailableCheckResponse;
 import com.wccg.well_c_code_git_backend.domain.user.model.User;
 import com.wccg.well_c_code_git_backend.domain.user.dto.service.request.UserSaveRequest;
 import com.wccg.well_c_code_git_backend.domain.user.repository.UserRepository;
-import com.wccg.well_c_code_git_backend.global.exception.exceptions.NicknameConflictException;
-import com.wccg.well_c_code_git_backend.global.exception.exceptions.NicknameLengthInvalidException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
-
-import static com.wccg.well_c_code_git_backend.domain.user.mapper.UserDtoMapper.toServiceNicknameAvailableCheckResponse;
 
 @Service
 @RequiredArgsConstructor
@@ -35,31 +29,11 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public Optional<User> getUserByGithubId(Long githubId){
+    public Optional<User> getUserByGithubId(Long githubId) {
         return userRepository.findByGithubId(githubId);
     }
 
-    public Optional<User> getUserById(Long id){
+    public Optional<User> getUserById(Long id) {
         return userRepository.findById(id);
-    }
-
-    public ServiceNicknameAvailableCheckResponse nicknameAvailableCheck(ServiceNicknameAvailableCheckRequest request) {
-        String nickname = request.getNickname();
-        nicknameLengthValidate(nickname);
-        nicknameConflictValidate(nickname);
-        return toServiceNicknameAvailableCheckResponse(true,nickname);
-    }
-
-    private void nicknameConflictValidate(String nickname) {
-        Optional<User> optionalUser = userRepository.findByNickname(nickname);
-        if(optionalUser.isPresent()){
-            throw new NicknameConflictException();
-        }
-    }
-
-    private void nicknameLengthValidate(String nickname) {
-        if(nickname.length() < 2 || nickname.length() > 12){
-            throw new NicknameLengthInvalidException();
-        }
     }
 }
